requirements.txt
gunicorn==20.1.0
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Flask-Bootstrap5==1.0.1
Flask-WTF==0.15.1
psycopg2-binary==2.9.1
python-dotenv==0.19.0
